+++ cybsi/api/auth/api_key.py
@@ -1,7 +1,10 @@
+import asyncio
+import json
 import logging
+import threading
 import uuid
 from datetime import datetime
-from typing import Generator, Iterable, List, Optional, Tuple
+from typing import AsyncGenerator, Generator, Iterable, List, Optional, Tuple, cast
 from urllib.parse import urljoin
 
 import httpx
@@ -25,7 +28,8 @@ logger = logging.getLogger(__name__)
 
 class APIKeyAuth(httpx.Auth):
     """Authomatically handles authentication
-    of :class:`~cybsi.api.CybsiClient` requests using API key.
+    of :class:`~cybsi.api.CybsiClient` or :class:`~cybsi.api.CybsiAsyncClient`
+    requests using API key.
 
     Args:
         api_url: Cybsi auth API URL. Usually equal to CybsiClient config API URL.
@@ -48,25 +52,45 @@ class APIKeyAuth(httpx.Auth):
     _get_token_path = "auth/token"
 
     def __init__(self, api_url: str, api_key: str, ssl_verify: bool = True):
+        # See https://www.python-httpx.org/advanced/#customizing-authentication
         self._api_key = api_key
-        self._token = None  # type: Optional[str]
         self._api_url = api_url
+        self._sync_lock = threading.RLock()
+        self._async_lock = asyncio.Lock()
+        self._token = ""
 
-    def auth_flow(
+    def sync_auth_flow(
         self, request: httpx.Request
     ) -> Generator[httpx.Request, httpx.Response, None]:
-        # See https://www.python-httpx.org/advanced/#customizing-authentication
-        if self._token is None:
-            token_response = yield self._build_token_request(request)
-            _raise_for_status(token_response)
 
-            token = TokenView(token_response.json())
-            self._token = f"{token.type.value} {token.access_token}"
+        with self._sync_lock:
+            if self._token:
+                request.headers["Authorization"] = self._token
+                response = yield request
 
-        request.headers["Authorization"] = self._token
-        yield request
+            if not self._token or response.status_code == 401:
+                token_response = yield self._build_token_request(request)
+                self._update_token(token_response, token_response.read())
 
-    def _build_token_request(self, req):
+            request.headers["Authorization"] = self._token
+            yield request
+
+    async def async_auth_flow(
+        self, request: httpx.Request
+    ) -> AsyncGenerator[httpx.Request, httpx.Response]:
+        async with self._async_lock:
+            if self._token:
+                request.headers["Authorization"] = self._token
+                response = yield request
+
+            if not self._token or response.status_code == 401:
+                token_response = yield self._build_token_request(request)
+                self._update_token(token_response, await token_response.aread())
+
+            request.headers["Authorization"] = self._token
+            yield request
+
+    def _build_token_request(self, req) -> httpx.Request:
         token_url = urljoin(self._api_url, self._get_token_path)
         headers = {
             "Accept": req.headers["Accept"],
@@ -78,6 +102,14 @@ class APIKeyAuth(httpx.Auth):
             params={"apiKey": self._api_key},
             headers=headers,
         )
+
+    def _update_token(
+        self, token_response: httpx.Response, token_response_content: bytes
+    ) -> None:
+        _raise_for_status(token_response)
+        token = TokenView(json.loads(token_response_content))
+
+        self._token = f"{token.type.value} {token.access_token}"
 
 
 class APIKeysAPI(BaseAPI):