+++ cybsi/api/internal/__init__.py
@@ -1,5 +1,6 @@
 from .base import (
     BaseAPI,
+    BaseAsyncAPI,
     JsonObject,
     JsonObjectForm,
     JsonObjectView,